Bekræftelse på fil sletning

+++ src/GUI/Controller/ProjectController.java
@@ -295,14 +295,30 @@ public class ProjectController extends BaseController {
 
         public void handleDeleteFile (ActionEvent actionEvent) throws Exception {
 
-            ProjectFiles fileToDelete = fileTable.getSelectionModel().getSelectedItem();
-            projectFilesModel.deleteFile(fileToDelete);
 
-            File file = new File(fileToDelete.getFilePath());
-            file.delete();
+            if (selectedfile != null) {
+                Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+                alert.setTitle("Delete File");
+                //alert.setContentText("Delete?");
+                ButtonType okButton = new ButtonType("Yes", ButtonBar.ButtonData.YES);
+                ButtonType noButton = new ButtonType("No", ButtonBar.ButtonData.NO);
+                alert.getButtonTypes().setAll(okButton, noButton);
+                alert.showAndWait().ifPresent(type -> {
+                    if (type == okButton) {
+                        try {
+                            projectFilesModel.deleteFile(selectedfile);
+                        } catch (Exception e) {
+                            throw new RuntimeException(e);
+                        }
+                        File file = new File(selectedfile.getFilePath());
+                        file.delete();
+                    }
 
+                });
+
+
+            }
         }
-
 
         public void handleOpenMainWindow (ActionEvent actionEvent) throws Exception {
             FXMLLoader loader = new FXMLLoader();
